feat: support using in CircleCI (#2)
I've added support for using this package in CircleCI, although it does require some extra config. We use ngrok to get around some of CircleCI's awkward networking config (docker daemon doesn't run on the same machine as the rest of the commands).

This also enables a proper CI setup for this package!